+++ virtualizarr/manifests/__init__.py
@@ -2,4 +2,6 @@
 # This is just to avoid conflicting with some type of file called manifest that .gitignore recommends ignoring.
 
 from .array import ManifestArray  # type: ignore # noqa
+from .group import ManifestGroup  # type: ignore # noqa
 from .manifest import ChunkEntry, ChunkManifest  # type: ignore # noqa
+from .store import ManifestStore  # type: ignore # noqa