+++ virtualizarr/tests/test_integration.py
@@ -58,7 +58,7 @@ def test_kerchunk_roundtrip_in_memory_no_concat(array_v3_metadata):
         pytest.param(
             5e7,
             ["lat", "lon", "time", "air"],
-            marks=pytest.mark.xfail(reason="scale factor encoding"),
+            marks=pytest.mark.skip(reason="slow"),
         ),
     ],
 )
@@ -75,11 +75,11 @@ def test_numpy_arrays_to_inlined_kerchunk_refs(
 
     # loading the variables should produce same result as inlining them using kerchunk
     with open_virtual_dataset(
-        netcdf4_file, loadable_variables=vars_to_inline, indexes={}, backend=hdf_backend
+        netcdf4_file, loadable_variables=vars_to_inline, backend=hdf_backend
     ) as vds:
         refs = vds.virtualize.to_kerchunk(format="dict")
 
-        # TODO I would just compare the entire dicts but kerchunk returns inconsistent results - see https://github.com/TomNicholas/VirtualiZarr/pull/73#issuecomment-2040931202
+        # TODO I would just compare the entire dicts but kerchunk returns inconsistent results - see https://github.com/zarr-developers/VirtualiZarr/pull/73#issuecomment-2040931202
         # assert refs == expected
         assert refs["refs"]["air/0.0.0"] == expected["refs"]["air/0.0.0"]
         assert refs["refs"]["lon/0"] == expected["refs"]["lon/0"]
@@ -149,8 +149,7 @@ class TestRoundtrip:
         with xr.tutorial.open_dataset("air_temperature", decode_times=False) as ds:
             # TODO: for now we will save as Zarr V3. Later we can parameterize it for V2.
             ds.to_zarr(air_zarr_path, zarr_format=3, consolidated=False)
-
-            with open_virtual_dataset(str(air_zarr_path), indexes={}) as vds:
+            with open_virtual_dataset(str(air_zarr_path)) as vds:
                 roundtrip = roundtrip_func(vds, tmp_path, decode_times=False)
 
                 # assert all_close to original dataset
@@ -175,13 +174,14 @@ class TestRoundtrip:
             ds.to_netcdf(air_nc_path)
 
             # use open_dataset_via_kerchunk to read it as references
-            with open_virtual_dataset(
-                str(air_nc_path), indexes={}, backend=hdf_backend
-            ) as vds:
+            with open_virtual_dataset(str(air_nc_path), backend=hdf_backend) as vds:
                 roundtrip = roundtrip_func(vds, tmp_path, decode_times=False)
 
                 # assert all_close to original dataset
                 xrt.assert_allclose(roundtrip, ds)
+
+                # TODO fails with ValueError: The truth value of an array with more than one element is ambiguous. Use a.any() or a.all()
+                # assert ds["air"].attrs == roundtrip["air"].attrs
 
                 # assert coordinate attributes are maintained
                 for coord in ds.coords:
@@ -215,13 +215,11 @@ class TestRoundtrip:
             with (
                 open_virtual_dataset(
                     str(air1_nc_path),
-                    indexes={},
                     loadable_variables=time_vars,
                     backend=hdf_backend,
                 ) as vds1,
                 open_virtual_dataset(
                     str(air2_nc_path),
-                    indexes={},
                     loadable_variables=time_vars,
                     backend=hdf_backend,
                 ) as vds2,
@@ -276,7 +274,7 @@ class TestRoundtrip:
         nc_path = tmp_path / "non_dim_coords.nc"
         ds.to_netcdf(nc_path)
 
-        with open_virtual_dataset(str(nc_path), indexes={}, backend=hdf_backend) as vds:
+        with open_virtual_dataset(str(nc_path), backend=hdf_backend) as vds:
             assert "lat" in vds.coords
             assert "coordinates" not in vds.attrs
 
@@ -332,14 +330,14 @@ def test_open_scalar_variable(tmp_path: Path, hdf_backend: type[VirtualBackend])
     ds = xr.Dataset(data_vars={"a": 0})
     ds.to_netcdf(nc_path)
 
-    with open_virtual_dataset(str(nc_path), indexes={}, backend=hdf_backend) as vds:
+    with open_virtual_dataset(str(nc_path), backend=hdf_backend) as vds:
         assert vds["a"].shape == ()
 
 
 @parametrize_over_hdf_backends
 class TestPathsToURIs:
     def test_convert_absolute_paths_to_uris(self, netcdf4_file, hdf_backend):
-        with open_virtual_dataset(netcdf4_file, indexes={}, backend=hdf_backend) as vds:
+        with open_virtual_dataset(netcdf4_file, backend=hdf_backend) as vds:
             expected_path = Path(netcdf4_file).as_uri()
             manifest = vds["air"].data.manifest.dict()
             path = manifest["0.0.0"]["path"]
@@ -349,9 +347,7 @@ class TestPathsToURIs:
     def test_convert_relative_paths_to_uris(self, netcdf4_file, hdf_backend):
         relative_path = relpath(netcdf4_file)
 
-        with open_virtual_dataset(
-            relative_path, indexes={}, backend=hdf_backend
-        ) as vds:
+        with open_virtual_dataset(relative_path, backend=hdf_backend) as vds:
             expected_path = Path(netcdf4_file).as_uri()
             manifest = vds["air"].data.manifest.dict()
             path = manifest["0.0.0"]["path"]