+++ virtualizarr/translators/kerchunk.py
@@ -95,9 +95,7 @@ def from_kerchunk_refs(decoded_arr_refs_zarray) -> "ArrayV3Metadata":
 
 def virtual_vars_and_metadata_from_kerchunk_refs(
     vds_refs: KerchunkStoreRefs,
-    loadable_variables,
-    drop_variables,
-    virtual_array_class=ManifestArray,
+    drop_variables: list[str] | None = None,
     fs_root: str | None = None,
 ) -> tuple[Mapping[str, Variable], dict[str, Any], list[str]]:
     """
@@ -105,6 +103,8 @@ def virtual_vars_and_metadata_from_kerchunk_refs(
 
     Parameters
     ----------
+    drop_variables
+        Variables in the file to not bother generating chunk metadata for.
     fs_root
         The root of the fsspec filesystem on which these references were generated.
         Required if any paths are relative in order to turn them into absolute paths (which virtualizarr requires).
@@ -112,8 +112,7 @@ def virtual_vars_and_metadata_from_kerchunk_refs(
 
     virtual_vars = virtual_vars_from_kerchunk_refs(
         vds_refs,
-        drop_variables=drop_variables + loadable_variables,
-        virtual_array_class=virtual_array_class,
+        drop_variables=drop_variables,
         fs_root=fs_root,
     )
     ds_attrs = fully_decode_arr_refs(vds_refs["refs"]).get(".zattrs", {})
@@ -164,7 +163,6 @@ def extract_group(vds_refs: KerchunkStoreRefs, group: str) -> KerchunkStoreRefs:
 def virtual_vars_from_kerchunk_refs(
     refs: KerchunkStoreRefs,
     drop_variables: list[str] | None = None,
-    virtual_array_class=ManifestArray,
     fs_root: str | None = None,
 ) -> dict[str, Variable]:
     """
@@ -174,9 +172,6 @@ def virtual_vars_from_kerchunk_refs(
     ----------
     drop_variables: list[str], default is None
         Variables in the file to drop before returning.
-    virtual_array_class
-        Virtual array class to use to represent the references to the chunks in each on-disk array.
-        Currently can only be ManifestArray, but once VirtualZarrArray is implemented the default should be changed to that.
     """
 
     var_names = find_var_names(refs)
@@ -187,9 +182,7 @@ def virtual_vars_from_kerchunk_refs(
     ]
 
     vars = {
-        var_name: variable_from_kerchunk_refs(
-            refs, var_name, virtual_array_class, fs_root=fs_root
-        )
+        var_name: variable_from_kerchunk_refs(refs, var_name, fs_root=fs_root)
         for var_name in var_names_to_keep
     }
     return vars
@@ -198,7 +191,6 @@ def virtual_vars_from_kerchunk_refs(
 def dataset_from_kerchunk_refs(
     refs: KerchunkStoreRefs,
     drop_variables: list[str] = [],
-    virtual_array_class: type = ManifestArray,
     indexes: MutableMapping[str, Index] | None = None,
     fs_root: str | None = None,
 ) -> Dataset:
@@ -207,14 +199,9 @@ def dataset_from_kerchunk_refs(
 
     drop_variables: list[str], default is None
         Variables in the file to drop before returning.
-    virtual_array_class
-        Virtual array class to use to represent the references to the chunks in each on-disk array.
-        Currently can only be ManifestArray, but once VirtualZarrArray is implemented the default should be changed to that.
     """
 
-    vars = virtual_vars_from_kerchunk_refs(
-        refs, drop_variables, virtual_array_class, fs_root=fs_root
-    )
+    vars = virtual_vars_from_kerchunk_refs(refs, drop_variables, fs_root=fs_root)
     ds_attrs = fully_decode_arr_refs(refs["refs"]).get(".zattrs", {})
     coord_names = ds_attrs.pop("coordinates", [])
 
@@ -235,7 +222,6 @@ def dataset_from_kerchunk_refs(
 def variable_from_kerchunk_refs(
     refs: KerchunkStoreRefs,
     var_name: str,
-    virtual_array_class,
     fs_root: str | None = None,
 ) -> Variable:
     """Create a single xarray Variable by reading specific keys of a kerchunk references dict."""
@@ -246,7 +232,7 @@ def variable_from_kerchunk_refs(
     dims = zattrs.pop("_ARRAY_DIMENSIONS")
     if chunk_dict:
         manifest = manifest_from_kerchunk_chunk_dict(chunk_dict, fs_root=fs_root)
-        varr = virtual_array_class(metadata=metadata, chunkmanifest=manifest)
+        varr = ManifestArray(metadata=metadata, chunkmanifest=manifest)
     elif len(metadata.shape) != 0:
         # empty variables don't have physical chunks, but zarray shows that the variable
         # is at least 1D
@@ -256,7 +242,7 @@ def variable_from_kerchunk_refs(
             metadata.chunks,
         )
         manifest = ChunkManifest(entries={}, shape=shape)
-        varr = virtual_array_class(metadata=metadata, chunkmanifest=manifest)
+        varr = ManifestArray(metadata=metadata, chunkmanifest=manifest)
     else:
         # This means we encountered a scalar variable of dimension 0,
         # very likely that it actually has no numeric value and its only purpose
@@ -276,7 +262,8 @@ def manifest_from_kerchunk_chunk_dict(
     for k, v in kerchunk_chunk_dict.items():
         if isinstance(v, (str, bytes)):
             raise NotImplementedError(
-                "Reading inlined reference data is currently not supported. [ToDo]"
+                "Reading inlined reference data is currently not supported."
+                "See https://github.com/zarr-developers/VirtualiZarr/issues/489",
             )
         elif not isinstance(v, (tuple, list)):
             raise TypeError(f"Unexpected type {type(v)} for chunk value: {v}")